Change PHP version to 8.0

+++ php8/ext/mysqli/tests/mysqli_pam_sha256.phpt
@@ -1,9 +1,10 @@
 --TEST--
 PAM: SHA-256
---EXTENSIONS--
-mysqli
 --SKIPIF--
 <?php
+require_once('skipif.inc');
+require_once('skipifconnectfailure.inc');
+
 ob_start();
 phpinfo(INFO_MODULES);
 $tmp = ob_get_contents();
@@ -11,9 +12,9 @@ ob_end_clean();
 if (!stristr($tmp, "auth_plugin_sha256_password"))
     die("skip SHA256 auth plugin not built-in to mysqlnd");
 
-require_once 'connect.inc';
-if (!$link = @my_mysqli_connect($host, $user, $passwd, $db, $port, $socket))
-    die(sprintf("skip Can't connect to MySQL Server - [%d] %s", mysqli_connect_errno(), mysqli_connect_error()));
+require_once('connect.inc');
+if (!$link = my_mysqli_connect($host, $user, $passwd, $db, $port, $socket))
+        die(printf("skip: [%d] %s\n", mysqli_connect_errno(), mysqli_connect_error()));
 
 if (mysqli_get_server_version($link) < 50606)
     die("skip: SHA-256 requires MySQL 5.6.6+");
@@ -44,8 +45,9 @@ if (strlen($row['Value']) < 100) {
     die(sprintf("skip Server misconfiguration? RSA pub key is suspicious, [%d] %s\n", $link->errno, $link->error));
 }
 
-// Ignore errors because this variable exists only in MySQL 5.6 and 5.7
-$link->query("SET @@session.old_passwords=2");
+if (!$link->query("SET @@session.old_passwords=2")) {
+    die(sprintf("skip Cannot set @@session.old_passwords=2 [%d] %s", $link->errno, $link->error));
+}
 
 $link->query('DROP USER shatest');
 $link->query("DROP USER shatest@localhost");
@@ -56,8 +58,8 @@ if (!$link->query('CREATE USER shatest@"%" IDENTIFIED WITH sha256_password') ||
     die(sprintf("skip CREATE USER failed [%d] %s", $link->errno, $link->error));
 }
 
-if (!$link->query('SET PASSWORD FOR shatest@"%" = "shatest"') ||
-    !$link->query('SET PASSWORD FOR shatest@"localhost" = "shatest"')) {
+if (!$link->query('SET PASSWORD FOR shatest@"%" = PASSWORD("shatest")') ||
+    !$link->query('SET PASSWORD FOR shatest@"localhost" = PASSWORD("shatest")')) {
     die(sprintf("skip SET PASSWORD failed [%d] %s", $link->errno, $link->error));
 }
 
@@ -73,11 +75,10 @@ if (!$link->query(sprintf("GRANT SELECT ON TABLE %s.test TO shatest@'%%'", $db))
 }
 
 $link->close();
-echo "nocache";
 ?>
 --FILE--
 <?php
-    require_once "connect.inc";
+    require_once("connect.inc");
 
     if (!$link = my_mysqli_connect($host, 'shatest', 'shatest', $db, $port, $socket)) {
         printf("[001] Cannot connect to the server using host=%s, user=%s, passwd=***, dbname=%s, port=%s, socket=%s\n",
@@ -103,7 +104,7 @@ echo "nocache";
 ?>
 --CLEAN--
 <?php
-    require_once "clean_table.inc";
+    require_once("clean_table.inc");
     $link->query('DROP USER shatest');
     $link->query('DROP USER shatest@localhost');
 ?>
