+++ php8/tests/basic/bug20539.phpt
@@ -1,8 +1,7 @@
 --TEST--
 Bug #20539 (PHP CLI Segmentation Fault)
---EXTENSIONS--
-session
 --SKIPIF--
+<?php if (!extension_loaded("session")) die("skip session extension not available"); ?>
 <?php if (file_exists(__DIR__. '/sess_' .session_id())) unlink(__DIR__. '/sess_' .session_id()); ?>
 --INI--
 session.auto_start=1
